requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
black==24.10.0
click==8.1.7
Django==5.1.4
django-cors-headers==4.6.0
django-extensions==3.2.3
django-ninja==1.3.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pydantic==2.10.3
pydantic_core==2.27.1
sqlparse==0.5.3
typing_extensions==4.12.2
